[refactor, fix] isSecret 항상 true로 넘어가는 에러 수정, 문의,리뷰모달 리팩토링
